+++ tests/test_forms.py
@@ -1,34 +1,18 @@
 from typing import Annotated
 
-from fastapi import Depends, Request
+from fastapi import Depends
 from fastapi.testclient import TestClient
 import air
 
-from pydantic import BaseModel, ValidationError
+from pydantic import BaseModel
 
 
 def test_simple_form():
     class CheeseModel(BaseModel):
-        name: str
-        age: int
+        name: str  # type: ignore [annotation-unchecked]
+        age: int  # type: ignore [annotation-unchecked]
 
-    class AirForm:
-        model = None
-        data = None
-        errors = None
-        is_valid = None
-
-        async def __call__(self, request: Request):
-            data = await request.form()
-            try:
-                self.data = self.model(**data)
-                self.is_valid = True
-            except ValidationError as e:
-                self.errors = e.errors()
-                self.is_valid = False
-            return self
-
-    class CheeseForm(AirForm):
+    class CheeseForm(air.AirForm):
         model = CheeseModel
 
     app = air.Air()
@@ -37,7 +21,7 @@ def test_simple_form():
     async def cheese_form(cheese: Annotated[CheeseForm, Depends(CheeseForm())]):
         if cheese.is_valid:
             return air.Html(air.H1(cheese.data.name))
-        return air.Html(air.H1(air.RawHTML(str(len(cheese.errors)))))
+        return air.Html(air.H1(air.RawHTML(str(len(cheese.errors)))))  # type: ignore [arg-type]
 
     client = TestClient(app)
